* server/fill_style.cpp: squash more non-standard functions (the 	  operations are on doubles anyway, so it doesn't involve any more 	  casts than before). Clean up logging. 	* server/parser/{BitmapMovieDefinition,movie_def_impl}.h: ditto. 	* server/as_value.cpp: use std::fmod, drop using namespace std;, 	  drop c_str() for logging, replace snprintf with boost::format 	  (will be slightly slower for those cases), drop snprintf ifdef. 	  A slightly faster alternative to boost::format that also avoids 	  using non-standard functions would be stringstreams, but it is 	  not as clear.

+++ server/parser/BitmapMovieDefinition.h
@@ -100,12 +100,12 @@ public:
 		return _version;
 	}
 
-	virtual float	get_width_pixels() const {
-		return ceilf(TWIPS_TO_PIXELS(_framesize.width()));
+	virtual float get_width_pixels() const {
+		return std::ceil(TWIPS_TO_PIXELS(_framesize.width()));
 	}
 
-	virtual float	get_height_pixels() const {
-		return ceilf(TWIPS_TO_PIXELS(_framesize.height()));
+	virtual float get_height_pixels() const {
+		return std::ceil(TWIPS_TO_PIXELS(_framesize.height()));
 	}
 
 	virtual size_t	get_frame_count() const {
